begin work on the 'universe' service
This will be its own beast, containing the workings, timings, and nature
of the game universe. This is separate from the controller, which
manages connections.

Introduces communication between the controller and the universe using
Thrift. Just some basic test stuff in there now, but it's enough to get
moving on some actual service communication.