send mobile movements to a reciprocal thrift connection
Having two TCP ports for thrift sucks, but I want to worry about async
RPC later and focus on overall architecture practices for now.

This may not work in the first place, as the mobileMovement call hangs
for a few minutes at a time after the first time it uses the call (which
is instant).